bump 0.8.1, fix key not reset when spawn()
This influences the behavior where field is defined in one place
and reused by spawn() in multiple ParamSet.

Before fixed, the following example would cause error:

    f = Field()

    class P0(ParamSet):
        key0 = f

    # after P0 is defined, f.key is `key0`

    class P1(ParamSet):
        key1 = f.spawn(null=True)

    # here the key of the new instance from f.spawn is still `key0`,
    # because ParamSet metaclass only assign key to Field instance
    # when key is None

This patch add `self.key = None` in spawn method, so that P1 can assign
the right key to field instance.

Note that if f is defined with `key='foo'`, when spawning, `key=` must
be passed again to keep as 'foo', otherwise it will be reset to None,
and waits for ParamSet to assign.